First commit!
This provides an outline for what will be implemented:

QAGD Diffuser is a structure for the main diffuser model.
QAGD Pipeline is a structure used for text generation.
QAGD Trainer is a structure used for training the model.
QAGD Schedulers is a class that contains utility methods.
logger helps logging in the console during training and running code.